chore: move folder to directly under components/*
