+++ src/components/nav-menu-button/nav-menu-button.mdx
@@ -1,6 +1,6 @@
 import { Canvas, Controls, Meta } from '@storybook/blocks'
 import * as NavMenuButtonStories from './nav-menu-button.stories'
-import { RenderHtmlMarkup } from '../../../../storybook/render-html-markup'
+import { RenderHtmlMarkup } from '../../storybook/render-html-markup'
 
 <Meta of={NavMenuButtonStories} />
